refactor kubeprovider
