+++ controllers/provisioners/provisioners.go
@@ -3,13 +3,13 @@ package provisioners
 import (
 	"github.com/go-logr/logr"
 	"github.com/keikoproj/instance-manager/api/v1alpha1"
-	"github.com/keikoproj/instance-manager/controllers/common"
 	awsprovider "github.com/keikoproj/instance-manager/controllers/providers/aws"
+	kubeprovider "github.com/keikoproj/instance-manager/controllers/providers/kubernetes"
 )
 
 type ProvisionerInput struct {
 	AwsWorker     awsprovider.AwsWorker
-	Kubernetes    common.KubernetesClientSet
+	Kubernetes    kubeprovider.KubernetesClientSet
 	InstanceGroup *v1alpha1.InstanceGroup
 	Configuration ProvisionerConfiguration
 	Log           logr.Logger
